requirements.txt
fastapi==0.100.0
uvicorn==0.22.0

SQLAlchemy==2.0.13
python-dotenv==1.0.0
mysql-connector-python

alembic==1.11.1

pydantic[email]
passlib[bcrypt]

pytest==7.4.0
httpx==0.24.1

python-jose==3.3.0

python-multipart==0.0.6

Jinja2==3.1.2